Split Readme file out into multiple files.
Split the data out from the README file into a CSV & head.md file.
The purpose of this is to enable future automated building of the
README.md.

The project is growing quickly in the scope of the data that needs to be
handled, such that manual manipulation is becoming very labor intensive.